lab1/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
djangorestframework==3.11.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
pipenv==2020.8.13
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
whitenoise==5.2.0